generative-ai/face-generation-comparison/requirements.txt
tensorflow==2.10.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pillow==10.0.1
numpy==1.24.3
requests==2.31.0
pydantic==2.4.2
protobuf==3.20.0
opencv-python==4.8.1